remove top level await
Removes the top level async function that had previously been wrapped around the transformed code. This ensures consistent behavior between node and lit-node, which became an issue after pull request #5 made it possible to load lit-node as a module instead of using it as a drop-in replacement for the node interpreter. This means top-level use of the await keyword is no longer allowed. Fixes #6.