move stack trace test
The test checking whether error locations are properly located in stack traces is part of a larger set of tests which are all contained in test.md, and it tests against its own location within that file, which means it is possible for the test to fail when other tests are edited if those edits result in changes to the line number of the stack trace test within the test.md file. This would be best handled be moving it into a discrete file which is unlikely to be edited at all, but so long as the tests are lumped together in a single monolithic test suite, it is at least useful to move the position of the stack trace up so edits to the other tests are less likely to affect its line number within the file.